add EntityManager

+++ src/main/java/web/controller/UserController.java
@@ -53,7 +53,7 @@ public class UserController {
     }
 
     @GetMapping("/delete/{id}")
-    public String deleteUser(@PathVariable("id") int id) {
+    public String deleteUser(@PathVariable("id") long id) {
         User user = userService.getById(id);
         userService.delete(user);
         return "redirect:/";
